Moved Spotify integration logic into backend
